Read the model's score as Float and validate with the 4B model
Metal has no double precision, so reading the mean log-probability as
Float64 raised an MLX error that trapped the app the first time a
remembered suggestion was validated. The 4B model replaces the 1B one
for suggestions, which was too inconsistent to trust.

+++ Sources/UttrflowLocalModel/MLXCandidateScorer.swift
@@ -122,6 +122,7 @@ public actor MLXCandidateScorer: CandidateScoring, CandidateGenerating {
         let taken = takeAlong(probabilities[(start - 1)..<(whole.count - 1)], targets, axis: 1)
         let mean = taken.mean()
         eval(mean)
-        return mean.item(Double.self)
+        // Read the scalar as Float, since Metal has no double precision and casting to Float64 errors.
+        return Double(mean.item(Float.self))
     }
 }
